Implement Basic Menu functionality
In order to improve my workflow this time around,
I had decided to follow the Unite 2017 lecture and
create an Event System using ScriptableObjects and UnityEvents.

In doing so, it:
Makes it much easier to visibly see GameObjects
interacting with each other through events in the inspector itself
Allows for events to be modified easily through drag and drop.
Abstract enough so that it can be used anywhere in the project.

I will use this system for now and see how it evolves in the
future (might switch back, but I do not know as of yet).