Improve core function logic of BasicMenu behavior
While the original Event System surrounding ScriptableObjects
was awesome to see in reality and fun to debug with, in reality
it was a huge pain in the ass to understand.

Due to it heavily relying on drag and drop components, it
made it very difficult to track. Furthermore, I felt
that while it made it easy to use, I wasn't practicing programming
due to the reliance on drag and drop (was missing out on the bigger
picture here).

As a result, I removed those elements and improved the logic
of the core functionality of BasicMenu instead.